usb-tools: serialc: Fixed average throughput calculation for >4GiB tests
The function that calculates the throughput took a size_t argument and
that wrapped around when the total length of the test exceeded 4 GiB,
causing the calculated average throughput to become very incorrect. This
patch corrects that by changing the type to uint64_t.

+++ serialc.c
@@ -273,7 +273,7 @@ static void release_interface(struct usb_serial_test *serial)
  * @usecs:	Number of microseconds taken
  * @size:	Size of data transfered
  */
-static float throughput(int64_t usecs, size_t size)
+static float throughput(int64_t usecs, uint64_t size)
 {
 	float bits = (float) size * 8.0;
 	float secs = (float) usecs / 1000000.0;
